requirements.txt
click==8.0.3
environs==9.3.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.0
psycopg2-binary==2.9.2
python-dotenv==0.19.2
Werkzeug==2.0.2
